refactor(global): move enums folder to top level
