+++ components/app/dao-proposals/DaoProposalsDropdowns.vue
@@ -17,7 +17,7 @@
 
 <script>
 import ButtonEmitWithDropdown from '@/components/common/ButtonEmitWithDropdown.vue';
-import CategoryEnum from '../../enums/Category.enum';
+import CategoryEnum from '../../../enums/Category.enum';
 import { CategoryMap } from '~/components/constants/CategoryMap.constant';
 
 export default {
