+++ components/app/jobs/JobStatus.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script lang="ts">
-import StatusEnum from '../../enums/Status.enum';
+import StatusEnum from '../../../enums/Status.enum';
 
 export default {
   data() {
