+++ components/app/projects/ProjectsCategories.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script>
-import CategoryEnum from '../../enums/Category.enum';
+import CategoryEnum from '../../../enums/Category.enum';
 
 export default {
   data() {
