+++ components/app/projects/ProjectsFilter.vue
@@ -17,7 +17,7 @@
 <script>
 import ProjectsDropdowns from './ProjectsDropdowns.vue';
 import AppSearchBar from '@/components/common/AppSearchbar.vue';
-import CategoryEnum from '@/components/enums/Category.enum';
+import CategoryEnum from '@/enums/Category.enum';
 import replaceQueryParams from '@/helpers/windowHistory.ts';
 
 export default {
